CORE-654 Deprecate console helper trait
The helper trait was only used once for `Console` and functionality has now been moved to `Console`.
The helper defined two abstract methods that are actually implemented by the parent console
class from Symfony. Our static code analysis showed that these abstract methods would actually
need to get implemented in deriving class (and not in the base class). Non of our console commands
actually does this so it's better to remove the trait since it's not reused anyways.